requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
peft>=0.5.0
accelerate>=0.23.0
bitsandbytes>=0.41.0
datasets>=2.14.0
trl>=0.7.2
scipy>=1.11.0
numpy>=1.24.0
scikit-learn>=1.2.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0
jsonlines>=3.1.0
typer>=0.9.0

# Distributed training
deepspeed>=0.10.0
mpi4py>=3.1.4

# Optimization
flash-attn>=2.3.0
optimum>=1.8.0
faiss-cpu>=1.7.4

# Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0
lm-eval-harness>=0.3.0
torchmetrics>=1.0.0

# Deployment
fastapi>=0.100.0
uvicorn>=0.22.0
gradio>=3.40.0
httpx>=0.24.0
sse-starlette>=1.6.0

# Notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
ipython>=8.15.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0
